created new files for components
