+++ src/components/DataJson.tsx
@@ -1,0 +1,8 @@
+interface IDataJson {
+  id: string;
+  img: string;
+  title: string;
+  price: number;
+  featured: boolean;
+}
+export default IDataJson;
