+++ src/components/SwitchView/index.js
@@ -1,0 +1,1 @@
+export {default} from './SwitchView.tsx';
